requirements.txt
Flask==3.0.3
Flask_MySQLdb==2.0.0
python-dotenv==1.0.1
